requirement.txt
pyenchant
bs4
pymysql
pydub
simpleaudio